+++ x-pack/platform/packages/shared/kbn-entities-schema/index.ts
@@ -5,10 +5,11 @@
  * 2.0.
  */
 
-export * from './src/schema/entity_definition';
-export * from './src/schema/entity';
-export * from './src/schema/common';
-export * from './src/schema/patterns';
+export * from './src/schema/v1/entity_definition';
+export * from './src/schema/v1/entity';
+export * from './src/schema/v2';
+export * from './src/schema/v1/common';
+export * from './src/schema/v1/patterns';
 export * from './src/rest_spec/create';
 export * from './src/rest_spec/delete';
 export * from './src/rest_spec/reset';
