+++ x-pack/platform/packages/shared/kbn-entities-schema/src/schema/v1/common.ts
